Implement Dapr API token authentication
Dapr API token authentication introduces an extra layer of security by having Dapr require every incoming gRPC and HTTP request for its APIs to include an authentication token before allowing the request to pass through.

The authentication token is generated using openssl and is stored as a k8s secret. The ingress Dapr sidecar is configured to have access only to that secret.